+++ lib/query_error.ex
@@ -1,0 +1,6 @@
+defmodule HtmlQuery.QueryError do
+  @moduledoc """
+  An exception raised when unable to find an HTML element.
+  """
+  defexception [:message]
+end